Travis: Test the Docker image
The build status of the docker image is added to the test to prevent
unnoticed failing docker image builds.

`-q` or quiet flag is raised for docker build because it'll flood
Travis build log and surpass the 4 MB limit, but if the docker build
fail it'll display the logs if it fails.